% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_sims.R
\name{multi_class_lr}
\alias{multi_class_lr}
\title{Multinomial Logistic Regression}
\usage{
multi_class_lr(n, mod1_size = 10, mod2_size = 10, rho = 0.8,
  beta = NULL)
}
\arguments{
\item{n}{Sample size.}

\item{mod1_size}{Size of first module.}

\item{mod2_size}{Size of second module.}

\item{rho}{Correlation of covariates.}

\item{beta}{A matrix of parameters.}
}
\value{
list with design matrix X, outcome y, and beta.
}
\description{
Function to generate multi-class data from a multinomial logistic
regression.  Assumes there are 5 classes.  Only supports two modules for now.
Currently this function is used for testing.
}
\note{
This work was partially funded by NSF IIS 1251151 and AMFAR 8721SC.
}
